Consolidate WooPlan and LegacyWooPlan
